small cleaning refactor
