+++ src/views/shop/ShopComponent.module.scss
@@ -13,7 +13,6 @@
 
     .shopView {
         @include boxShadowSm;
-
         background: rgba($color-light-grey, 0.65);
         border-radius: 0.62rem;
         padding: 1.25rem;
